@winged 2019-06-21
WIP because cache is a simple dict that doesn't evict old data yet